+++ util.go
@@ -9,24 +9,25 @@ import (
 )
 
 // JSON HTTP Request をするだけのラッパー
-func PostRequest(reqUrl string, reqBody interface{}) ([]byte, error) {
-	_, err := url.ParseRequestURI(reqUrl)
+func PostRequest(reqURL string, reqBody interface{}) ([]byte, error) {
+	_, err := url.ParseRequestURI(reqURL)
 	if err != nil {
 		return nil, err
 	}
-	reqJson, err := json.Marshal(reqBody)
+	reqJSON, err := json.Marshal(reqBody)
 	if err != nil {
 		return nil, err
 	}
 	req, err := http.NewRequest(
 		"POST",
-		reqUrl,
-		bytes.NewBuffer([]byte(reqJson)),
+		reqURL,
+		bytes.NewBuffer([]byte(reqJSON)),
 	)
 	if err != nil {
 		logger.Info("post request error, error=", err)
 		return nil, err
 	}
+	req.Header.Set("Content-Type", "application/json")
 	client := &http.Client{}
 	resp, err := client.Do(req)
 	if err != nil {