+++ CHANGES.md
@@ -9,8 +9,21 @@
 - FIX
     - バグ修正
 
+
 ## develop
+- [ADD] CI の go を 1.13 に上げる
+- [UPDATE] @kdxu 推奨の go version を 1.13 にあげる
+- [UPDATE] @kdxu 先に入室しているユーザーがいる場合 isExistUser をaccept時にtrue にして返す
+- [FIX] @kdxu roomId か clientId が空文字列の場合は reject するように修正する
 - [FIX] @kdxu -c での config ファイル指定が効いていなかったのを修正する
+
+## 19.08.0
+
+2019-08-16
+
+- [UPDATE] `/ws` エンドポイントと同様のものを `/signaling` エンドポイントとして追加する
+- [UPDATE] ayame register 時に key も送信できるようにする
+- [UPDATE] auth webhook の返り値に iceServers があれば返却するようにする
 
 ## 19.07.1
 - [CHANGE] @kdxu サンプルを ayame-web-sdk を用いたものに置き換える